HTTP requests timeout in 120s
By default, 'requests', which means we will still hang in cases where
the HTTP request for some reason failed. Instead of wait indefinitely on
the response (and hang instead of return), we now specify a timeout
duration to prevent these things from happening.

Currently, the timeout is set to 120s, and we are not paying any
particular attention to the result (ie, currently we are not catching
and dealing with the Timeout exception).