raise the memory size
The feed we're using is currently providing a lot more items (100!),
which means that the previous memory size of 72 is not enough to
accomodate it.

Since there is no way to properly cope with a feed bigger than the
memory, not only this commit raises the memory size, we now also emit
a warning if that sittuation comes up.